Move network partition functions out of q_config.c
These were the only bits that didn't have anything to do with setting up
the configuration or dumping it, but rather with looking up network
partitions in an existing configuration.
