+++ src/core/ddsi/include/dds/ddsi/q_misc.h
@@ -40,6 +40,18 @@ DDS_EXPORT int guid_prefix_eq (const ddsi_guid_prefix_t *a, const ddsi_guid_pref
 DDS_EXPORT int guid_eq (const struct ddsi_guid *a, const struct ddsi_guid *b);
 DDS_EXPORT int ddsi2_patmatch (const char *pat, const char *str);
 
+#ifdef DDSI_INCLUDE_NETWORK_PARTITIONS
+struct ddsi_config;
+struct ddsi_config_partitionmapping_listelem;
+struct ddsi_config_partitionmapping_listelem *find_partitionmapping (const struct ddsi_config *cfg, const char *partition, const char *topic);
+int is_ignored_partition (const struct ddsi_config *cfg, const char *partition, const char *topic);
+#endif
+#ifdef DDSI_INCLUDE_NETWORK_CHANNELS
+struct ddsi_config;
+struct ddsi_config_channel_listelem;
+struct ddsi_config_channel_listelem *find_channel (const struct config *cfg, nn_transport_priority_qospolicy_t transport_priority);
+#endif
+
 #if defined (__cplusplus)
 }
 #endif
